USB: EHCI: Add MSM UICC controller driver
EHCI compliant Full-Speed only controller is used for UICC connectivity
on MSM8939. The controller is configured in 3-wire serial mode and there
is no separate PHY. The TLMM block in MSM converts the controller serial
signals into DP and DM signals.

Manage the clock and gpio as part of root hub PM. The remote wakeup support
is not added in this patch. It can be implemented by configuration the
DP/DM gpio line as an interrupt.

[rcky844]: Integrate b833231226b2 ("USB: UICC: set bus->skip_resume to true
           for UICC drivers")

CRs-Fixed: 671394
Change-Id: I8c14230eb567b6ffef6b5a9e8dc312a197e0811e
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>