USB: ice40-hcd: Import ICE40 FPGA based SPI to Inter-Chip USB host driver
Taken from msm-3.10 at commit b833231226b2 ("USB: UICC: set
bus->skip_resume to true for UICC drivers").

The ICE40 Host controller driver (ice40-hcd) is registered as a SPI
protocol driver and interacts with the SPI subsystem on one side and
interacts with the USB core on the other side. It supports only 1
Full-Speed port. The bridge chip does not maintain any port states.
Root hub port logic is simulated in the software. Control and Bulk
transfers are supported.

This driver is used by msm8916, so pull in this driver.

Signed-off-by: Ricky Cheung <rcheung844@gmail.com>
Change-Id: Ic32e5bc32b4e27c0d77c026353ea7d69c83fb8ba