ASoC: wcd: Add option for disabling micbias connected to regulator
This config has been added to resolve a redefinition error for the
enums involving on demand supply with msm8x16-wcd. Only compile into
the wcd9335 codec driver when this option is not set.

Since all legacy msm8916/39 devices does not use this option, select
it to resolve a build error.

Change-Id: I003a4337b7d027e8ad8b1e07e5a6bddd0d17aed0
Signed-off-by: Ricky Cheung <rcheung844@gmail.com>